correct examples
